# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Morex API",
        "description": "API pour l'application mobile Morex - Gestion des finances personnelles",
        "contact": {
            "name": "Morex Support",
            "email": "support@morex.app"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/stats/dashboard": {
            "get": {
                "tags": [
                    "Stats"
                ],
                "summary": "Dashboard principal - Statistiques globales",
                "description": "Retourne toutes les statistiques nécessaires pour le dashboard de l'app mobile",
                "operationId": "09dba8957203724ab1f443d629227d91",
                "responses": {
                    "200": {
                        "description": "Statistiques du dashboard",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "period": {
                                            "properties": {
                                                "month": {
                                                    "type": "string",
                                                    "example": "January 2026"
                                                },
                                                "start": {
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "end": {
                                                    "type": "string",
                                                    "format": "date"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "balance": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 1500000
                                                },
                                                "formatted": {
                                                    "type": "string",
                                                    "example": "1 500 000 FCFA"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "monthly": {
                                            "properties": {
                                                "income": {
                                                    "type": "integer"
                                                },
                                                "expense": {
                                                    "type": "integer"
                                                },
                                                "balance": {
                                                    "type": "integer"
                                                },
                                                "savings_rate": {
                                                    "type": "number"
                                                },
                                                "formatted_income": {
                                                    "type": "string"
                                                },
                                                "formatted_expense": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "goals": {
                                            "properties": {
                                                "active_count": {
                                                    "type": "integer"
                                                },
                                                "total_target": {
                                                    "type": "integer"
                                                },
                                                "total_current": {
                                                    "type": "integer"
                                                },
                                                "progress": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "debts": {
                                            "properties": {
                                                "total_debt": {
                                                    "type": "integer"
                                                },
                                                "total_credit": {
                                                    "type": "integer"
                                                },
                                                "net_position": {
                                                    "type": "integer"
                                                },
                                                "overdue_count": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "alerts": {
                                            "properties": {
                                                "exceeded_budgets": {
                                                    "type": "integer"
                                                },
                                                "overdue_debts": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "top_expense_categories": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "color": {
                                                        "type": "string"
                                                    },
                                                    "total": {
                                                        "type": "integer"
                                                    },
                                                    "formatted_total": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/stats/monthly": {
            "get": {
                "tags": [
                    "Stats"
                ],
                "summary": "Statistiques mensuelles détaillées",
                "operationId": "72360010e136f8136f3ff49c44abe0a7",
                "parameters": [
                    {
                        "name": "month",
                        "in": "query",
                        "description": "Format YYYY-MM",
                        "schema": {
                            "type": "string",
                            "example": "2026-01"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statistiques du mois",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "month": {
                                            "type": "string",
                                            "example": "2026-01"
                                        },
                                        "period": {
                                            "type": "object"
                                        },
                                        "summary": {
                                            "properties": {
                                                "income": {
                                                    "type": "integer"
                                                },
                                                "expense": {
                                                    "type": "integer"
                                                },
                                                "balance": {
                                                    "type": "integer"
                                                },
                                                "transaction_count": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "income_by_category": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "expense_by_category": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "daily": {
                                            "description": "Transactions par jour",
                                            "type": "object"
                                        },
                                        "by_account": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification via token Sanctum. Utilisez le token reçu lors du login.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}